+++ spring-functionaltest-domain/src/main/java/jp/co/ntt/fw/spring/functionaltest/domain/service/dam3mb3/OrderMB3ServiceImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2018 NTT Corporation.
+ * Copyright(c) 2014 NTT Corporation.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -9,9 +9,9 @@
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
  */
 package jp.co.ntt.fw.spring.functionaltest.domain.service.dam3mb3;
 
@@ -74,8 +74,8 @@ public class OrderMB3ServiceImpl implements OrderMB3Service {
         long total = orderMB3Repository.countById();
         List<OrderMB3> todos;
         if (0 < total) {
-            RowBounds rowBounds = new RowBounds(pageable.getOffset(), pageable
-                    .getPageSize());
+            RowBounds rowBounds = new RowBounds((int) pageable
+                    .getOffset(), pageable.getPageSize());
             todos = orderMB3Repository.findPageMyBatis3(rowBounds);
         } else {
             todos = Collections.emptyList();
@@ -90,8 +90,8 @@ public class OrderMB3ServiceImpl implements OrderMB3Service {
         long total = orderMB3Repository.countById();
         List<OrderMB3> todos;
         if (0 < total) {
-            RowBounds rowBounds = new RowBounds(pageable.getOffset(), pageable
-                    .getPageSize());
+            RowBounds rowBounds = new RowBounds((int) pageable
+                    .getOffset(), pageable.getPageSize());
             todos = orderMB3Repository.findPageMyBatis3Scroll(rowBounds);
         } else {
             todos = Collections.emptyList();
